Remove `drupal/devel` from `require-dev`
Drupal modules should always be in the `require` section, and not treated as dev dependencies even if they're a development related module.

In the past I had recommended that we put `devel` in the `require-dev` section, but that was a mistake. Since `config_split` determines what is enabled or not, having it missing depending on how you run `composer install` doesn't make sense.